chore: Skeleton out the actor
