+++ backend/src/manager/mod.rs
@@ -12,10 +12,10 @@ use embassy_container_init::{ProcessGroupId, SignalGroupParams};
 use helpers::RpcClient;
 use nix::sys::signal::Signal;
 use patch_db::DbHandle;
-use sqlx::{Executor, Postgres};
+use sqlx::Executor;
 use tokio::sync::watch::error::RecvError;
 use tokio::sync::watch::{channel, Receiver, Sender};
-use tokio::sync::{oneshot, Notify, RwLock};
+use tokio::sync::{oneshot, Notify};
 use torut::onion::TorSecretKeyV3;
 use tracing::instrument;
 
@@ -27,123 +27,16 @@ use crate::procedure::docker::{DockerContainer, DockerProcedure, LongRunning};
 #[cfg(feature = "js_engine")]
 use crate::procedure::js_scripts::JsProcedure;
 use crate::procedure::{NoOutput, PackageProcedure, ProcedureName};
-use crate::s9pk::manifest::{Manifest, PackageId};
-use crate::util::{ApplyRef, Container, NonDetachingJoinHandle, Version};
+use crate::s9pk::manifest::Manifest;
+use crate::util::{ApplyRef, Container, NonDetachingJoinHandle};
 use crate::Error;
 
 pub mod health;
+pub mod manager_map;
 mod sync;
 
 pub const HEALTH_CHECK_COOLDOWN_SECONDS: u64 = 15;
 pub const HEALTH_CHECK_GRACE_PERIOD_SECONDS: u64 = 5;
-
-#[derive(Default)]
-pub struct ManagerMap(RwLock<BTreeMap<(PackageId, Version), Arc<Manager>>>);
-impl ManagerMap {
-    #[instrument(skip(self, ctx, db, secrets))]
-    pub async fn init<Db: DbHandle, Ex>(
-        &self,
-        ctx: &RpcContext,
-        db: &mut Db,
-        secrets: &mut Ex,
-    ) -> Result<(), Error>
-    where
-        for<'a> &'a mut Ex: Executor<'a, Database = Postgres>,
-    {
-        let mut res = BTreeMap::new();
-        for package in crate::db::DatabaseModel::new()
-            .package_data()
-            .keys(db)
-            .await?
-        {
-            let man: Manifest = if let Some(manifest) = crate::db::DatabaseModel::new()
-                .package_data()
-                .idx_model(&package)
-                .and_then(|pkg| pkg.installed())
-                .map(|m| m.manifest())
-                .get(db)
-                .await?
-                .to_owned()
-            {
-                manifest
-            } else {
-                continue;
-            };
-
-            let tor_keys = man.interfaces.tor_keys(secrets, &package).await?;
-            res.insert(
-                (package, man.version.clone()),
-                Arc::new(Manager::create(ctx.clone(), man, tor_keys).await?),
-            );
-        }
-        *self.0.write().await = res;
-        Ok(())
-    }
-
-    #[instrument(skip(self, ctx))]
-    pub async fn add(
-        &self,
-        ctx: RpcContext,
-        manifest: Manifest,
-        tor_keys: BTreeMap<InterfaceId, TorSecretKeyV3>,
-    ) -> Result<(), Error> {
-        let mut lock = self.0.write().await;
-        let id = (manifest.id.clone(), manifest.version.clone());
-        if let Some(man) = lock.remove(&id) {
-            if !man.thread.is_empty().await {
-                man.exit().await?;
-            }
-        }
-        lock.insert(
-            id,
-            Arc::new(Manager::create(ctx, manifest, tor_keys).await?),
-        );
-        Ok(())
-    }
-
-    #[instrument(skip(self))]
-    pub async fn remove(&self, id: &(PackageId, Version)) {
-        if let Some(man) = self.0.write().await.remove(id) {
-            if let Err(e) = man.exit().await {
-                tracing::error!("Error shutting down manager: {}", e);
-                tracing::debug!("{:?}", e);
-            }
-        }
-    }
-
-    #[instrument(skip(self))]
-    pub async fn empty(&self) -> Result<(), Error> {
-        let res =
-            futures::future::join_all(std::mem::take(&mut *self.0.write().await).into_iter().map(
-                |((id, version), man)| async move {
-                    tracing::debug!("Manager for {}@{} shutting down", id, version);
-                    man.exit().await?;
-                    tracing::debug!("Manager for {}@{} is shutdown", id, version);
-                    if let Err(e) = Arc::try_unwrap(man) {
-                        tracing::trace!(
-                            "Manager for {}@{} still has {} other open references",
-                            id,
-                            version,
-                            Arc::strong_count(&e) - 1
-                        );
-                    }
-                    Ok::<_, Error>(())
-                },
-            ))
-            .await;
-        res.into_iter().fold(Ok(()), |res, x| match (res, x) {
-            (Ok(()), x) => x,
-            (Err(e), Ok(())) => Err(e),
-            (Err(e1), Err(e2)) => Err(Error::new(eyre!("{}, {}", e1.source, e2.source), e1.kind)),
-        })
-    }
-
-    #[instrument(skip(self))]
-    pub async fn get(&self, id: &(PackageId, Version)) -> Option<Arc<Manager>> {
-        self.0.read().await.get(id).cloned()
-    }
-}
-
 pub struct Manager {
     shared: Arc<ManagerSharedState>,
     thread: Container<NonDetachingJoinHandle<()>>,
