+++ backend/src/util/mod.rs
@@ -23,6 +23,7 @@ use tracing::instrument;
 
 use crate::shutdown::Shutdown;
 use crate::{Error, ErrorKind, ResultExt as _};
+pub mod actor;
 pub mod config;
 pub mod http_reader;
 pub mod io;
